perf(webapp): load the ffmpeg command body on first use
`index.ts` imported the whole `ffmpeg` command — argv grammar, camera
capture, WORKERFS staging, both engines — into the kernel worker's
boot-critical graph, and the engine work in this series had grown that
graph by 4.1 kB. `ffmpeg-command.ts` is now the same first-use stub
`ffprobe` and `pdftotext` already use; the body moves to `ffmpeg/run.ts`
unchanged. Worker eager payload: -18 kB vs main instead of +4.1 kB.
